fix: remove svc only from env with `svc delete --env`
Right now, when a customer runs `svc delete --env e`
copilot will delete the svc from the entire application,
including the ECR repo and the SSM parameter. We only
want to do that if the customer entierly deletes the service.

This change updates svc delete to only remove the env/svc
stack when an environment flag is left in.

Future improvements can be to empty the repo if there
are no more deployed services. I don't think this
is a big deal though.

fixes #1126